+++ src/solver.h
@@ -7,7 +7,7 @@
 #include "full.h"
 #include "tucker.h"
 
-const double PI = acos(-1.0); // TODO
+const REAL PI = acos(-1.0); // TODO
 
 enum AlgoritmParts {
 	RECONSTRUCTION,
@@ -21,56 +21,56 @@ class GasParams {
 public:
 	// Default constructor
 
-	double Na = 6.02214129e+23; // Avogadro constant
-	double kB = 1.381e-23; // Boltzmann constant, J / K
-	double Ru = 8.3144598; // Universal gas constant
+	REAL Na = 6.02214129e+23; // Avogadro constant
+	REAL kB = 1.381e-23; // Boltzmann constant, J / K
+	REAL Ru = 8.3144598; // Universal gas constant
 
-	double Mol = 40e-3; // = Mol
-	double Rg = Ru / Mol; // = self.Ru  / self.Mol  # J / (kg * K)
-	double m = Mol / Na; // # kg
+	REAL Mol = 40e-3; // = Mol
+	REAL Rg = Ru / Mol; // = self.Ru  / self.Mol  # J / (kg * K)
+	REAL m = Mol / Na; // # kg
 
-	double g = 5.0 / 3.0; // # specific heat ratio
-	double d = 3.418e-10; // # diameter of molecule
+	REAL g = 5.0 / 3.0; // # specific heat ratio
+	REAL d = 3.418e-10; // # diameter of molecule
 
-	double Pr = 2.0 / 3.0;
+	REAL Pr = 2.0 / 3.0;
 
-	double C = 144.4;
-	double T_0 = 273.11;
-	double mu_0 = 2.125e-5;
+	REAL C = 144.4;
+	REAL T_0 = 273.11;
+	REAL mu_0 = 2.125e-5;
 
-	double mu_suth(double T) const;
-	double mu(double T) const;
+	REAL mu_suth(REAL T) const;
+	REAL mu(REAL T) const;
 };
 
 template <class Tensor>
 class VelocityGrid {
 public:
 	// Constructor
-	VelocityGrid(int nvx_, int nvy_, int nvz_, double *vx__, double *vy__, double *vz__);
+	VelocityGrid(int nvx_, int nvy_, int nvz_, REAL *vx__, REAL *vy__, REAL *vz__);
 	// Destructor
 	~VelocityGrid();
 
-	double *vx_;
-	double *vy_;
-	double *vz_;
+	REAL *vx_;
+	REAL *vy_;
+	REAL *vz_;
 
 	int nvx, nvy, nvz;
 	int nv;
 
-	double hvx, hvy, hvz;
-	double hv3;
+	REAL hvx, hvy, hvz;
+	REAL hv3;
 
-	double *vx;
-	double *vy;
-	double *vz;
+	REAL *vx;
+	REAL *vy;
+	REAL *vz;
 
-	double *zerox;
-	double *zeroy;
-	double *zeroz;
+	REAL *zerox;
+	REAL *zeroy;
+	REAL *zeroz;
 
-	double *onesx;
-	double *onesy;
-	double *onesz;
+	REAL *onesx;
+	REAL *onesy;
+	REAL *onesz;
 
 	Tensor vx_t;
 	Tensor vy_t;
@@ -97,19 +97,19 @@ public:
 
 	bcType type;
 
-	virtual Tensor applyBC(double x, double y, double z, 
+	virtual Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) = 0;
+			REAL tol) = 0;
 };
 
 template <class Tensor>
 class BCSYMMETRYX : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
+			REAL tol) override {
 				return reflect(f, 'X');
 	}
 };
@@ -117,10 +117,10 @@ public:
 template <class Tensor>
 class BCSYMMETRYY : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
+			REAL tol) override {
 				return reflect(f, 'Y');
 	}
 };
@@ -128,10 +128,10 @@ public:
 template <class Tensor>
 class BCSYMMETRYZ : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
+			REAL tol) override {
 				return reflect(f, 'Z');
 	}
 };
@@ -139,10 +139,10 @@ public:
 template <class Tensor>
 class BCINLET : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
+			REAL tol) override {
 				return Tensor(BoundaryCondition<Tensor>::bcData);
 	}
 };
@@ -150,10 +150,10 @@ public:
 template <class Tensor>
 class BCOUTLET : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
+			REAL tol) override {
 				return Tensor(BoundaryCondition<Tensor>::bcData);
 	}
 };
@@ -161,12 +161,12 @@ public:
 template <class Tensor>
 class BCWALL : public BoundaryCondition<Tensor> {
 public:
-	Tensor applyBC(double x, double y, double z, 
+	Tensor applyBC(REAL x, REAL y, REAL z, 
 			const Tensor& f, 
 			const Tensor& vn, const Tensor& vn_abs, 
-			double tol) override {
-				double Ni = BoundaryCondition<Tensor>::v->hv3 * (round_t(0.5 * f * (vn + vn_abs), tol, 1e+6)).sum();
-				double Nr = BoundaryCondition<Tensor>::v->hv3 * (round_t(0.5 * BoundaryCondition<Tensor>::bcData * (vn - vn_abs), tol, 1e+6)).sum();
+			REAL tol) override {
+				REAL Ni = BoundaryCondition<Tensor>::v->hv3 * (round_t(0.5 * f * (vn + vn_abs), tol, 1e+6)).sum();
+				REAL Nr = BoundaryCondition<Tensor>::v->hv3 * (round_t(0.5 * BoundaryCondition<Tensor>::bcData * (vn - vn_abs), tol, 1e+6)).sum();
 				return -(Ni / Nr) * BoundaryCondition<Tensor>::bcData;
 	}
 };
@@ -178,7 +178,7 @@ public:
 	std::shared_ptr < VelocityGrid<Tensor> > v;
 
 	std::vector < Tensor > initData;
-	Tensor getInit(double x, double y, double z,
+	Tensor getInit(REAL x, REAL y, REAL z,
 			const std::vector<Tensor>& initData);
 
 	std::vector < int > bcTags;
@@ -189,8 +189,8 @@ public:
 struct Config {
 	bool isImplicit = false;
 
-	double CFL = 0.5;
-	double tol = 1e-7;
+	REAL CFL = 0.5;
+	REAL tol = 1e-7;
 
 	std::string initType = "default";
 	std::string initFilename = "";
@@ -230,8 +230,8 @@ public:
 	std::vector < Tensor > vn_abs;
 	Tensor vn_abs_r1;
 
-	double h;
-	double tau;
+	REAL h;
+	REAL tau;
 
 	std::vector < Tensor > diag;
 	std::vector < Tensor > diag_r1;
@@ -244,29 +244,29 @@ public:
 	std::vector < Tensor > df;
 
 	// Arrays for macroparameters
-	std::vector < double > n;
-	std::vector < double > rho;
-	std::vector < double > ux, uy, uz;
-	std::vector < double > p;
-	std::vector < double > T;
-	std::vector < double > nu;
-	std::vector < double > rank;
-	std::vector < double > compression;
-	std::vector < std::vector < double > > data;
+	std::vector < REAL > n;
+	std::vector < REAL > rho;
+	std::vector < REAL > ux, uy, uz;
+	std::vector < REAL > p;
+	std::vector < REAL > T;
+	std::vector < REAL > nu;
+	std::vector < REAL > rank;
+	std::vector < REAL > compression;
+	std::vector < std::vector < REAL > > data;
 	std::map<AlgoritmParts, std::vector<double>> timings;
 
 	std::vector < BoundaryCondition<Tensor> *> bcList; // TODO make shared
 
-	std::vector <double> comp_macro_params(const Tensor& f);
-	Tensor comp_j(const std::vector <double>& params, const Tensor& f);
+	std::vector <REAL> comp_macro_params(const Tensor& f);
+	Tensor comp_j(const std::vector <REAL>& params, const Tensor& f);
 	void write_wall_params();
 
 	int it;
-	std::vector <double> frob_norm_iter;
+	std::vector <REAL> frob_norm_iter;
 
 	void plot_residual();
 	void create_res();
-	void update_res(double frob_norm);
+	void update_res(REAL frob_norm);
 
 	void save_tec();
 	void save_macro();
@@ -282,13 +282,13 @@ public:
 };
 
 template <class Tensor>
-double *f_maxwell(std::shared_ptr < VelocityGrid<Tensor> > v,
-		double n, double ux, double uy, double uz,
-		double T, double Rg);
+REAL *f_maxwell(std::shared_ptr < VelocityGrid<Tensor> > v,
+		REAL n, REAL ux, REAL uy, REAL uz,
+		REAL T, REAL Rg);
 
 template <class Tensor>
 Tensor f_maxwell_t(std::shared_ptr < VelocityGrid<Tensor> > v,
-		double n, double ux, double uy, double uz,
-		double T, double Rg);
+		REAL n, REAL ux, REAL uy, REAL uz,
+		REAL T, REAL Rg);
 
 #endif /* SOLVER_H_ */
